requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
asttokens
async-timeout
attrs
bokeh
certifi
charset-normalizer
click
comm
contourpy
cycler
debugpy
decorator
docker-pycreds
einops
exceptiongroup
executing
filelock
fonttools
frozenlist
fsspec
gitdb
GitPython
idna
imageio
ipykernel
ipython
ipywidgets
jedi
Jinja2
joblib
jupyter_client
jupyter_core
jupyterlab_widgets
kiwisolver
kornia
kornia_rs
lazy_loader
lightning-utilities
llvmlite
MarkupSafe
matplotlib
matplotlib-inline
mpmath
multidict
nest-asyncio
networkx
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
opencv-python
opencv-python-headless
packaging
pandas
parso
pexpect
pillow
pip
platformdirs
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pure_eval
py-cpuinfo
pycocotools
Pygments
pynndescent
pyparsing
PyQt5
PyQt5-Qt5
PyQt5_sip
python-dateutil
pytorch-lightning
pytz
PyYAML
pyzmq
requests
scikit-image
scikit-learn
scipy
seaborn
sentry-sdk
setproctitle
setuptools
six
smmap
stack-data
sympy
threadpoolctl
tifffile
torch
torchmetrics
torchvision
tornado
tqdm
traitlets
triton
typing_extensions
tzdata
ultralytics
ultralytics-thop
umap-learn
urllib3
wandb
wcwidth
widgetsnbextension
xyzservices
yarl
